+++ .gitignore
@@ -4,6 +4,11 @@ build/
 !**/src/main/**/build/
 !**/src/test/**/build/
 
+target/
+.bsp/
+examples/scala-examples/lib/*
+
+
 ### IntelliJ IDEA ###
 .idea/*
 *.iws
