--- examples/build.gradle
@@ -1,48 +1,0 @@
-plugins {
-    id 'java'
-    id 'org.jetbrains.kotlin.jvm' version '1.9.0-Beta'
-}
-
-group 'com.edgedb'
-version '1.0-SNAPSHOT'
-
-sourceSets.main {
-    java.srcDirs("src/main/java", "src/main/kotlin")
-}
-
-repositories {
-    mavenCentral()
-    maven  {
-        url "https://oss.sonatype.org/content/repositories/snapshots"
-    }
-}
-
-dependencies {
-    implementation project(path: ':src:driver')
-    implementation 'ch.qos.logback:logback-classic:1.4.5'
-    implementation 'ch.qos.logback:logback-core:1.4.5'
-    implementation 'org.jooq:joou:0.9.4'
-    implementation 'org.reflections:reflections:0.10.2'
-    implementation 'com.fasterxml.jackson.core:jackson-databind:2.15.2'
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.1'
-
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
-    testImplementation 'junit:junit:4.13.2'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
-}
-
-test {
-    useJUnitPlatform()
-}
-compileKotlin {
-    destinationDirectory = compileJava.destinationDirectory
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
-}
-compileTestKotlin {
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
-}
