+++ examples/kotlin-examples/src/main/kotlin/com/edgedb/examples/Example.kt
@@ -1,0 +1,7 @@
+package com.edgedb.examples
+
+import com.edgedb.driver.EdgeDBClient
+
+interface Example {
+    suspend fun runAsync(client: EdgeDBClient)
+}
