+++ examples/scala-examples/project/build.properties
@@ -1,0 +1,1 @@
+sbt.version = 1.9.0
