+++ examples/scala-examples/project/plugins.sbt
@@ -1,0 +1,1 @@
+addSbtPlugin("org.jetbrains.scala" % "sbt-ide-settings" % "1.1.1")
