--- examples/src/main/java/com/edgedb/examples/Example.java
@@ -1,9 +1,0 @@
-package com.edgedb.examples;
-
-import com.edgedb.driver.EdgeDBClient;
-
-import java.util.concurrent.CompletionStage;
-
-public interface Example {
-    CompletionStage<Void> run(EdgeDBClient client);
-}
