--- examples/src/main/java/com/edgedb/examples/LinkProperties.java
@@ -1,38 +1,0 @@
-package com.edgedb.examples;
-
-import com.edgedb.driver.EdgeDBClient;
-import com.edgedb.driver.annotations.EdgeDBType;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.Collection;
-import java.util.concurrent.CompletionStage;
-
-public final class LinkProperties implements Example {
-    private static final Logger logger = LoggerFactory.getLogger(LinkProperties.class);
-
-    @EdgeDBType
-    public static final class Person {
-        public String name;
-        public Long age;
-        public Person bestFriend;
-        public Collection<Person> friends;
-    }
-
-    private static final String INSERT_QUERY =
-            "with a := (insert Person { name := 'Person A', age := 20 } unless conflict on .name)," +
-            "b := (insert Person { name := 'Person B', age := 21 } unless conflict on .name)," +
-            "c := (insert Person { name := 'Person C', age := 22, friends := b } unless conflict on .name)" +
-            "insert Person { name := 'Person D', age := 23, friends := { a, b, c }, best_friend := c } unless conflict on .name";
-
-    @Override
-    public CompletionStage<Void> run(EdgeDBClient client) {
-        return client.execute(INSERT_QUERY)
-                .thenCompose(v ->
-                        client.queryRequiredSingle(
-                                Person.class,
-                                "select Person { name, age, friends: { name, age, friends }, best_friend: { name, age, friends } } filter .name = 'Person D'"
-                        ))
-                .thenAccept(result -> logger.info("Person with links: {}", result));
-    }
-}
