--- examples/src/main/java/com/edgedb/examples/QueryResults.java
@@ -1,32 +1,0 @@
-package com.edgedb.examples;
-
-import com.edgedb.driver.EdgeDBClient;
-import com.edgedb.driver.annotations.EdgeDBType;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.concurrent.CompletionStage;
-
-public final class QueryResults implements Example {
-    private static final Logger logger = LoggerFactory.getLogger(QueryResults.class);
-
-    @EdgeDBType
-    public static final class Person {
-        public String name;
-        public Long age;
-
-        public void setName(String name) {
-            this.name = name;
-        }
-        public void setAge(Long age) {
-            this.age = age;
-        }
-    }
-
-    @Override
-    public CompletionStage<Void> run(EdgeDBClient client) {
-        return client.execute("insert Person { name := 'Example', age := 1234 } unless conflict on .name")
-                .thenCompose(v -> client.queryRequiredSingle(Person.class, "select Person { name, age } filter .name = 'Example'"))
-                .thenAccept(result -> logger.info("Person returned from query: {}", result));
-    }
-}
