--- examples/src/main/kotlin/com/edgedb/examples/KotlinExample.kt
@@ -1,7 +1,0 @@
-package com.edgedb.examples
-
-import com.edgedb.driver.EdgeDBClient
-
-interface KotlinExample {
-    suspend fun runAsync(client: EdgeDBClient)
-}
