--- examples/src/main/kotlin/com/edgedb/examples/KotlinGlobalsAndConfig.kt
@@ -1,31 +1,0 @@
-package com.edgedb.examples
-
-import com.edgedb.driver.EdgeDBClient
-import kotlinx.coroutines.future.await
-import org.slf4j.LoggerFactory
-import java.time.Duration
-import java.util.*
-
-class KotlinGlobalsAndConfig : KotlinExample {
-    companion object {
-        private val logger = LoggerFactory.getLogger(KotlinExample::class.java)!!
-    }
-
-    override suspend fun runAsync(client: EdgeDBClient) {
-        val configuredClient = client
-                .withConfig { config -> config
-                    .withIdleTransactionTimeout(Duration.ZERO)
-                    .applyAccessPolicies(true)
-                }
-                .withGlobals(mapOf(
-                    "current_user_id" to UUID.randomUUID()
-                ))
-
-        val currentUserId = configuredClient.queryRequiredSingle(
-                UUID::class.java,
-                "SELECT GLOBAL current_user_id"
-        ).await()
-
-        logger.info("Current user ID: {}", currentUserId)
-    }
-}
