--- examples/src/main/resources/logback.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!-- Log level: OFF, ERROR, WARN, INFO, DEBUG, TRACE, ALL -->
-
-<configuration>
-  <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-    <layout class="ch.qos.logback.classic.PatternLayout">
-      <Pattern>%d{HH:mm:ss} [%thread] %class{0} %M : %msg%n</Pattern>
-    </layout>
-  </appender>
-  <appender name="APPEND" class="ch.qos.logback.classic.AsyncAppender">
-    <appender-ref ref="STDOUT" />
-    <includeCallerData>true</includeCallerData>
-  </appender>
-
-  <logger name="nio2.ssl.SSLEngineAutomat" level="ALL" additivity="false">
-    <appender-ref ref="STDOUT" />
-  </logger>
-  <logger name="nio2.ssl.SSLAsynchronousSocketChannel" level="ALL" additivity="false">
-    <appender-ref ref="STDOUT" />
-  </logger>
-  <logger name="nio2.ssl.SSLAsynchronousSocketChannelLayer" level="ALL" additivity="false">
-    <appender-ref ref="STDOUT" />
-  </logger>
-  <logger name="nio2.ssl.SSLAsynchronousSocketChannelImpl" level="ALL" additivity="false">
-    <appender-ref ref="STDOUT" />
-  </logger>
- <logger name="nio2.ssl.SSLAsynchronousChannelGroup" level="ALL" additivity="false">
-    <appender-ref ref="STDOUT" />
-  </logger>
-  <root level="ALL">
-    <appender-ref ref="APPEND" />
-  </root>
-</configuration>
